requirements.txt
# FastAPI and web framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# CORS and middleware
python-dotenv==1.0.0

# LangChain and RAG
langchain==0.1.0
langchain-community==0.0.13
langchain-core==0.1.0

# Document processing
pypdf==3.17.4
python-docx==1.1.0

# Text processing and embeddings
scikit-learn==1.3.2
numpy==1.24.3

# Language detection
langdetect==1.0.9
langid==1.1.6

# HTTP client for Ollama
requests==2.31.0

# Pydantic for data validation
pydantic==2.5.0

# Optional: Redis (if you want to add it back later)
# redis==5.0.1

# Development tools (optional)
# pytest==7.4.3
# black==23.11.0